requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
pandas_ta==0.3.14b0
python-dotenv==1.0.1
loguru==0.7.2
requests==2.32.3
pytz==2024.2
python-telegram-bot==21.6
